+++ logic/l_types.nim
@@ -9,10 +9,26 @@ import times, l_log, strutils
 
 type
   TWeight* {.exportc.} = object of TObject
+    # Fields serialized to the database.
     Fid: int ## Internal db identifier of the object.
     Fdate: TTime ## When was the object recorded.
     Fweight: float ## Weight, according to type.
     Ftyp: Weight_type ## Type of the weight.
+
+    Fday_date: string ## date field in YYYYMMDD format for day comparisons. \
+    ##
+    ## May be nil, use the accessor to automatically recalculate it. The date
+    ## is calculated using the local time by default. This may be problematic
+    ## for unit testing, requiring a future l_db.USE_GMT_TIME hack.
+    alternating_day*: bool ## Marks entries as odd/even days for colouring.
+
+    changes_day*: bool ## \
+    ## Set to true if the immediate previous entry is for a different day.
+    ##
+    ## A group of three entries with the same day will have the same vale for
+    ## `alternating_day`, but only the first will have true in the
+    ## `changes_day` field.
+
   PWeight* {.exportc.} = ref TWeight
 
   Weight_type* = enum kilograms = 0, pounds = 1 ## \
@@ -79,10 +95,13 @@ proc id*(s: TWeight): int =
 
 proc `date=`*(s: var TWeight; value: TTime) =
   ## Setter for the date attribute.
+  ##
+  ## Also cleans the day_date field by forcing it to nil.
   s.Fdate = value
+  s.Fday_date = nil
 
 proc date*(s: PWeight): TTime {.raises:[].} =
-  ## Getter for the date attribute.
+  ## Getter for the Fdate attribute.
   ##
   ## Returns zero if the weight is invalid.
   if not s.isNil:
@@ -96,7 +115,7 @@ proc typ*(s: TWeight): Weight_type =
   ## Getter for the type of weight.
   return s.Ftyp
 
-proc short*(x: Weight_type): string =
+proc short*(x: Weight_type): string not nil =
   case x:
   of kilograms: result = kg_str
   of pounds: result = lb_str
@@ -123,6 +142,10 @@ proc format_iso*(info: TTimeInfo): string {.raises: [EInvalidValue].} =
   result = info.format("yyyy-MM-dd:HH-mm-ss")
 
 
+proc safe_str*(s: string): string {.raises: [].} =
+  ## Returns the string or "(nil)" if the variable is not initialized.
+  result = if s.isNil: "(nil)" else: s
+
 proc `$`*(x: PWeight): string {.raises: [].} =
   ## Debugging/logging helper of weights.
   if x.isNil:
@@ -136,8 +159,8 @@ proc `$`*(x: PWeight): string {.raises: [].} =
 
   # Build debug string.
   result = "PWeight {id:" & $(x[].id) & ", weight:" &
-    formatFloat(x.weight, ffDecimal, 4) & ", typ:" & x[].typ_str &
-    ", date:" & d & "}"
+    formatFloat(x.Fweight, ffDecimal, 4) & ", typ:" & x[].typ_str &
+    ", date:" & d & ", day_date:" & safe_str(x[].Fday_date) & "}"
 
 
 proc `==`(x, y: PWeight): bool =
@@ -158,3 +181,19 @@ proc exists_value*(s: seq[PWeight], x: PWeight): bool =
     if y == x:
       result = true
       return
+
+proc day_date*(w: var TWeight): string {.raises: [EInvalidValue].} =
+  ## Returns the day_date field of PWeight, which may need recalculation.
+  if w.Fday_date.isNil:
+    let info = w.Fdate.get_local_time
+    w.Fday_date = info.format("yyyyMMdd")
+  assert w.Fday_date.len > 0
+  return w.Fday_date
+
+
+proc day_date*(w: PWeight): string =
+  ## Returns the day_date field of PWeight, which may need recalculation.
+  ##
+  ## Returns the empty string if the weight pointer is nil.
+  if w.isNil: return ""
+  return day_date(w[])