+++ objc_interface/formatters.h
@@ -1,0 +1,7 @@
+#import "n_global.h"
+
+NSString *format_date(TWeight *weight);
+NSString *format_nsdate(NSDate *date);
+
+NSAttributedString *format_shadowed_date(TWeight *weight,
+    id font, id normal_color, id shadowed_color);