+++ docs/changes.rst
@@ -2,6 +2,20 @@
 Seohtracker logic changes log
 =============================
 
+v5, 2014-04-01
+--------------
+
+* `Added generic objc interface
+  <https://github.com/gradha/seohtracker-logic/issues/4>`_.
+* `Made objc interface format dates using relative format
+  <https://github.com/gradha/seohtracker-logic/issues/7>`_.
+* `Added alternating bool to loaded entries
+  <https://github.com/gradha/seohtracker-logic/issues/9>`_.
+* `Marked get_weight_string() as returning not nil
+  <https://github.com/gradha/seohtracker-logic/issues/6>`_.
+* `Factored common code from client nakefiles
+  <https://github.com/gradha/seohtracker-logic/issues/8>`_.
+
 v4, 2014-03-06
 --------------